+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 20250925 - v0.8.1
+
+- **[update]：** 重命名 rosdistro submodule，与官方仓库保持一致
+- **[update]：** 为 rosdistro submodule 增加 `shallow = true` 参数，提升仓库克隆速度
+- **[update]：** 移除 `run.sh` 中冗余的 `PrepareRosdistro` 操作，提升 ROS 安装速度
+
 ## 20250922 - v0.8
 
 - **[new]：** 添加对 Ubuntu 22.04（Jammy）下 ROS 2 Humble 的支持；